projects_manager/requirements.txt
Django>=4.0
environ==1.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
django-environ==0.10.0
